chore: update dependencies & manage breaks
