+++ .eslintignore
@@ -1,2 +1,3 @@
 svelte.config.js
 mdsvex.config.js
+vite.config.js
